backend/app/translate/formatters/pdf.py
"""
PDF 文档格式处理器
支持 .pdf 文件的翻译
"""
import uuid
from pathlib import Path
from typing import Callable, Optional

from . import BaseFormatter
from ...config import settings


class PDFFormatter(BaseFormatter):
    """
    PDF 文档处理器

    PDF 翻译比较复杂，需要使用专门的库
    这里提供基础实现，生产环境建议使用 babeldoc 或专业服务
    """

    def translate(
        self,
        source_path: str,
        target_lang: str,
        ai_translator,
        progress_callback: Optional[Callable[[int, int], None]] = None
    ) -> str:
        """
        翻译 PDF 文档

        注意：完整实现需要安装额外依赖（如 PyMuPDF、pdfplumber 等）
        这里提供简化框架
        """
        # TODO: 实现 PDF 翻译逻辑
        # 生产环境建议使用 babeldoc 或其他专业 PDF 翻译库

        # 简化实现：提取文本，翻译后生成新的 PDF
        # 注意：这种方法会丢失原始 PDF 的格式和布局

        # 提取文本（使用 PyMuPDF）
        try:
            import fitz  # PyMuPDF
        except ImportError:
            raise ImportError(
                "PDF 处理需要安装 PyMuPDF: pip install PyMuPDF"
            )

        doc = fitz.open(source_path)
        texts = []

        for page in doc:
            text = page.get_text()
            if text.strip():
                texts.append(text)

        doc.close()

        # 翻译
        translated_texts = []
        total = len(texts)

        for i, text in enumerate(texts):
            translated = ai_translator.translate_text(text, target_lang)
            translated_texts.append(translated)

            if progress_callback:
                progress_callback(i + 1, total)

        # 生成结果（简化版：保存为文本文件）
        # 实际应用中应该使用 reportlab 等库生成 PDF
        result_path = self._generate_result_path(source_path, ext='.txt')

        with open(result_path, 'w', encoding='utf-8') as f:
            f.write('\n\n'.join(translated_texts))

        return result_path

    def _generate_result_path(self, source_path: str, ext: str = None) -> str:
        """生成结果文件路径"""
        source = Path(source_path)
        if ext is None:
            ext = source.suffix
        filename = f"{source.stem}_translated_{uuid.uuid4().hex[:8]}{ext}"
        return str(settings.TRANSLATE_DIR / filename)
