backend/requirements.txt
# DocTranslator 后端依赖清单（FastAPI 版本 - 兼容 Python 3.13）

# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7

# ORM 扩展 - 使用更新的版本以支持 Python 3.13
pydantic>=2.10
pydantic-settings>=2.6.1
email-validator>=2.1.1

# 认证
python-jose[cryptography]==4.0.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# 数据库迁移
alembic==1.13.1

# 文档处理
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# AI 翻译
openai==1.6.1

# Redis（可选，用于缓存和任务队列）
redis==5.0.1

# HTTP 客户端
httpx==0.26.0
